requirements.txt
python-dotenv
langchain==0.2.1
langchain-community==0.2.1
openai
langchain-openai
chromadb==0.5.0
pypdf==4.2.0
pydantic